wonder3d_oa/multi_view_generation/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
diffusers[torch]==0.31.0
xformers
bitsandbytes==0.35.4
decord==0.6.0
pytorch-lightning<2
omegaconf==2.2.3
nerfacc==0.3.3
trimesh==3.9.8
pyhocon==0.3.57
icecream==2.1.0
PyMCubes==0.1.2
accelerate
modelcards
einops
ftfy
piq
matplotlib
opencv-python
imageio
imageio-ffmpeg
scipy
pyransac3d
torch_efficient_distloss
tensorboard
rembg
segment_anything
gradio==3.50.2
